重构APIC和VMPL模块，使用MSR寄存器替代内联汇编，简化MSR读写操作，改进CR2和FSBASE的读取逻辑

+++ src/vc/ghcb.rs
@@ -25,7 +25,7 @@ use libc::MAP_FAILED;
 use crate::CreateType;
 use crate::mm::pgtable_va_to_pa;
 
-pub const MSR_AMD64_SEV_ES_GHCB: u64 = 0xc0010130;
+pub const MSR_AMD64_SEV_ES_GHCB: u32 = 0xc0010130;
 pub const GHCB_MMAP_BASE: VirtAddr = VirtAddr::new(0);
 
 /// 1
